fluid sim code
This is the code and all accompanying files to open this in visual studio. Note that if you want to move over the code to another IDE you need to create a new project in that IDE and copy over all the code. Also note in order to run this code you need to include GLFW, glew, glm and stb for the code to compile correctly :) Other than that it's just to run the code!